pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_analyzer"]

[project]
name = "mcp-youtube-analyzer"
version = "0.1.0"
description = "MCP server for analyzing YouTube videos — transcripts, frame extraction, and metadata"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Saphrogi" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "youtube-transcript-api>=1.2.0",
    "yt-dlp>=2024.0.0",
    "av>=12.0.0",
    "Pillow>=10.0.0",
    "scenedetect>=0.6.0",
]

[project.optional-dependencies]
whisper = ["faster-whisper>=1.0.0"]

[project.urls]
Homepage = "https://github.com/TommasoBagassi/mcp-youtube-analyzer"
Repository = "https://github.com/TommasoBagassi/mcp-youtube-analyzer"
Issues = "https://github.com/TommasoBagassi/mcp-youtube-analyzer/issues"

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
mcp-youtube-analyzer = "youtube_analyzer.server:main"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests requiring network access (deselect with '-m \"not integration\"')",
]
